Move inactiveCurrencies sync-schema to separate file
