+++ workers/loc.api/sync/schema/sync-schema/inactive-currencies.js
@@ -1,0 +1,18 @@
+'use strict'
+
+const TABLES_NAMES = require('../tables-names')
+const ALLOWED_COLLS = require('../allowed.colls')
+const COLLS_TYPES = require('../colls.types')
+
+const { getModelOf } = require('../models')
+
+module.exports = {
+  name: ALLOWED_COLLS.INACTIVE_CURRENCIES,
+  maxLimit: 10000,
+  projection: 'pairs',
+  sort: [['pairs', 1]],
+  hasNewData: false,
+  isSyncRequiredAtLeastOnce: true,
+  type: COLLS_TYPES.PUBLIC_UPDATABLE_ARRAY,
+  model: getModelOf(TABLES_NAMES.INACTIVE_CURRENCIES)
+}
