+++ workers/loc.api/sync/schema/sync-schema/index.js
@@ -27,6 +27,7 @@ const tickersHistory = require('./tickers-history')
 const wallets = require('./wallets')
 const symbols = require('./symbols')
 const mapSymbols = require('./map-symbols')
+const inactiveCurrencies = require('./inactive-currencies')
 
 const _methodCollMap = new Map([
   [SYNC_API_METHODS.LEDGERS, ledgers],
@@ -48,19 +49,7 @@ const _methodCollMap = new Map([
   [SYNC_API_METHODS.WALLETS, wallets],
   [SYNC_API_METHODS.SYMBOLS, symbols],
   [SYNC_API_METHODS.MAP_SYMBOLS, mapSymbols],
-  [
-    SYNC_API_METHODS.INACTIVE_CURRENCIES,
-    {
-      name: ALLOWED_COLLS.INACTIVE_CURRENCIES,
-      maxLimit: 10000,
-      projection: 'pairs',
-      sort: [['pairs', 1]],
-      hasNewData: false,
-      isSyncRequiredAtLeastOnce: true,
-      type: COLLS_TYPES.PUBLIC_UPDATABLE_ARRAY,
-      model: getModelOf(TABLES_NAMES.INACTIVE_CURRENCIES)
-    }
-  ],
+  [SYNC_API_METHODS.INACTIVE_CURRENCIES, inactiveCurrencies],
   [
     SYNC_API_METHODS.INACTIVE_SYMBOLS,
     {
